refactor: extract TestCredentialsApplier
We are going to introduce a basic auth credentials applier. For our tests we need this TestCredentialsApplier, so
instead of copying it, it's nicer to extract it and reuse the code in both test classes.
